Fix/news blog update (#129)
* fix: fix update status

* fix: return created time after fetch api

+++ internal/application/service/news_service_impl.go
@@ -146,6 +146,7 @@ func (newsService *NewsService) GetAdminNews(newsID uint) (newsdto.AdminNewsResp
 
 	return newsdto.AdminNewsResponse{
 		ID:          news.ID,
+		CreatedAt:   news.CreatedAt,
 		Title:       news.Title,
 		Content:     news.Content,
 		Description: news.Description,
@@ -177,6 +178,7 @@ func (newsService *NewsService) GetPublicNews(newsID uint) (newsdto.PublicNewsRe
 
 	return newsdto.PublicNewsResponse{
 		ID:          news.ID,
+		CreatedAt:   news.CreatedAt,
 		Title:       news.Title,
 		Content:     news.Content,
 		Description: news.Description,
@@ -213,6 +215,7 @@ func (newsService *NewsService) GetAdminNewsList(request newsdto.GetAdminNewsLis
 
 		newsResponse[i] = newsdto.AdminNewsResponse{
 			ID:          eachNews.ID,
+			CreatedAt:   eachNews.CreatedAt,
 			Title:       eachNews.Title,
 			Content:     eachNews.Content,
 			Description: eachNews.Description,
@@ -254,6 +257,7 @@ func (newsService *NewsService) GetPublicNewsList(request newsdto.GetPublicNewsL
 
 		newsResponse[i] = newsdto.PublicNewsResponse{
 			ID:          eachNews.ID,
+			CreatedAt:   eachNews.CreatedAt,
 			Title:       eachNews.Title,
 			Content:     eachNews.Content,
 			Description: eachNews.Description,
@@ -362,30 +366,28 @@ func (newsService *NewsService) EditNews(request newsdto.EditNewsRequest) error 
 		news.Description = *request.Description
 	}
 
-	newStatus := newsService.mapToOperationalStatuses(request.Status)
-	if err := newsService.checkStatusConflict(newStatus, news.Status); err != nil {
-		return err
+	if request.Status != nil && *request.Status != uint(news.Status) {
+		news.Status = newsService.mapToOperationalStatuses(*request.Status)
 	}
-	news.Status = newStatus
 
+	oldCoverPath := news.CoverImage
 	if request.CoverImage != nil {
-		newCoverPath := newsService.constants.S3BucketPath.GetNewsCoverImagePath(news.ID, request.CoverImage.Filename)
-		if news.CoverImage != newCoverPath {
-			oldCoverPath := news.CoverImage
-			news.CoverImage = newCoverPath
-			if err := newsService.s3Storage.UploadObject(enum.NewsMedia, news.CoverImage, request.CoverImage); err != nil {
-				return err
-			}
-			if oldCoverPath != "" {
-				if err := newsService.s3Storage.DeleteObject(enum.NewsMedia, oldCoverPath); err != nil {
-					return err
-				}
-			}
+		news.CoverImage = newsService.constants.S3BucketPath.GetNewsCoverImagePath(news.ID, request.CoverImage.Filename)
+		if err := newsService.s3Storage.UploadObject(enum.NewsMedia, news.CoverImage, request.CoverImage); err != nil {
+			return err
 		}
 	}
 
-	err = newsService.newsRepository.UpdateNews(newsService.db, news)
-	return err
+	if err = newsService.newsRepository.UpdateNews(newsService.db, news); err != nil {
+		return err
+	}
+
+	if request.CoverImage != nil && oldCoverPath != "" {
+		if err := newsService.s3Storage.DeleteObject(enum.NewsMedia, oldCoverPath); err != nil {
+			return err
+		}
+	}
+	return nil
 }
 
 func (newsService *NewsService) UpdateNewsStatus(request newsdto.EditNewsStatusRequest) error {
